+++ include/nstd/os/windows/def.h
@@ -1,0 +1,8 @@
+#ifndef NSTD_OS_WINDOWS_DEF_H_INCLUDED
+#define NSTD_OS_WINDOWS_DEF_H_INCLUDED
+#include "../../core/def.h"
+
+/// Represents a Windows HANDLE.
+typedef NSTDISize NSTDOSWindowsHandle;
+
+#endif