+++ src/os/windows.rs
@@ -1,4 +1,6 @@
 #[cfg(feature = "nstd_os_alloc")]
 pub mod alloc;
+#[cfg(feature = "nstd_os_def")]
+pub mod def;
 #[cfg(feature = "nstd_os_io")]
 pub mod io;