+++ src/os/windows/def.rs
@@ -1,0 +1,2 @@
+/// Represents a Windows HANDLE.
+pub type NSTDOSWindowsHandle = isize;